decouple admin hook presentation tests from database
