+++ server/utils/adminHooks/envelope.ts
@@ -1,7 +1,6 @@
 import type { User } from '../../drizzle/schema';
 import type { SiteConfig } from '../../types/config';
 import { getGlobalConfig } from '../config';
-import { labelForAdminHookEvent } from './localization';
 import type { AdminHookActor, AdminHookEnvelope } from './types';
 
 function getSiteInfo() {
@@ -33,42 +32,6 @@ function toActor(user: Partial<User> | null | undefined): AdminHookActor {
 function previewText(value: string, maxLength = 160) {
   const normalized = value.replace(/\s+/g, ' ').trim();
   return normalized.length > maxLength ? `${normalized.slice(0, maxLength)}...` : normalized;
-}
-
-export function titleForEnvelope(envelope: AdminHookEnvelope) {
-  return `${envelope.site.name}: ${labelForAdminHookEvent(
-    envelope.event,
-    envelope.site.defaultLanguage
-  )}`;
-}
-
-export function bodyForEnvelope(envelope: AdminHookEnvelope) {
-  if (envelope.event === 'user.registered') {
-    return [envelope.user?.username, envelope.user?.nickname].filter(Boolean).join(' / ');
-  }
-  return [
-    envelope.note?.author?.username || envelope.actor.username,
-    envelope.note?.title,
-    envelope.note?.contentPreview,
-  ]
-    .filter(Boolean)
-    .join(' / ');
-}
-
-export function urlForEnvelope(envelope: AdminHookEnvelope) {
-  if (envelope.note?.url) return envelope.note.url;
-
-  if (
-    envelope.event === 'user.registered' &&
-    envelope.user?.username &&
-    envelope.site.frontendUrl
-  ) {
-    return `${envelope.site.frontendUrl.replace(/\/+$/, '')}/${encodeURIComponent(
-      envelope.user.username
-    )}`;
-  }
-
-  return envelope.site.frontendUrl;
 }
 
 export function buildUserRegisteredEnvelope(user: Partial<User>): AdminHookEnvelope {
